requirements.txt
annotated-types==0.7.0
APScheduler==3.11.2
greenlet==3.4.0
playwright==1.58.0
pydantic==2.13.2
pydantic_core==2.46.2
pyee==13.0.1
python-dotenv==1.2.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
